// src/app/classes/media-handling.enum.ts
export enum MediaHandlingEnum {
  outspoken_unflappable = 'Outspoken, Unflappable',
  outspoken_shortTempered_confrontational = 'Outspoken, Short-Tempered, Confrontational',
  outspoken_shortTempered = 'Outspoken, Short-Tempered',
  outspoken_volatile_confrontational = 'Outspoken, Volatile, Confrontational',
  outspoken_volatile = 'Outspoken, Volatile',
  outspoken_confrontational = 'Outspoken, Confrontational',
  outspoken = 'Outspoken',
  evasive_unflappable = 'Evasive, Unflappable',
  evasive_shortTempered_confrontational = 'Evasive, Short-Tempered, Confrontational',
  evasive_shortTempered = 'Evasive, Short-Tempered',
  evasive_volatile_confrontational = 'Evasive, Volatile, Confrontational',
  evasive_volatile = 'Evasive, Volatile',
  evasive_confrontational = 'Evasive, Confrontational',
  evasive_reserved = 'Evasive, Reserved',
  evasive = 'Evasive',
  unflappable = 'Unflappable',
  shortTempered_confrontational = 'Short-Tempered, Confrontational',
  shortTempered = 'Short-Tempered',
  volatile_confrontational = 'Volatile, Confrontational',
  volatile = 'Volatile',
  confrontational = 'Confrontational',
  reserved = 'Reserved',
  levelHeaded = 'Level-Headed',
  mediaFrendly_unflappable = 'Media-Frendly, Unflappable',
  mediaFrendly_shortTempered_confrontational = 'Media-Frendly, Short-Tempered, Confrontational',
  mediaFrendly_shortTempered = 'Media-Frendly, Short-Tempered',
  mediaFrendly_volatile_confrontational = 'Media-Frendly, Volatile, Confrontational',
  mediaFrendly_volatile = 'Media-Frendly, Volatile',
  mediaFrendly_confrontational = 'Media-Frendly, Confrontational',
  mediaFrendly_reserved = 'Media-Frendly, Reserved',
  mediaFrendly = 'Media-Frendly',
}